htmlchecker: Move _get_pattern() helper to Checker class
This will be used to replace some very similar code in gitchecker.
